# backend/openapi-schema.yml
components:
  schemas:
    ApiDocumentWithTasks:
      properties:
        changed_at:
          title: Changed At
          type: string
        created_at:
          title: Created At
          type: string
        id:
          format: uuid
          title: Id
          type: string
        media_files:
          items:
            $ref: '#/components/schemas/RemoteDocumentMedia'
          title: Media Files
          type: array
        name:
          title: Name
          type: string
        tasks:
          items:
            $ref: '#/components/schemas/TaskResponse'
          title: Tasks
          type: array
      required:
      - id
      - name
      - created_at
      - changed_at
      - media_files
      - tasks
      title: ApiDocumentWithTasks
      type: object
...
    Body_create_document_api_v1_documents__post:
      properties:
        file:
          contentMediaType: application/octet-stream
          title: File
          type: string
        language:
          title: Language
          type: string
        model:
          $ref: '#/components/schemas/TranscriptionModel'
        name:
          title: Name
          type: string
        number_of_speakers:
          anyOf:
          - minimum: 0.0
            type: integer
          - type: 'null'
          title: Number Of Speakers
      required:
      - name
      - model
      - language
      - file
      title: Body_create_document_api_v1_documents__post
      type: object
    Body_import_document_api_v1_documents_import__post:
      properties:
        media_file:
          contentMediaType: application/octet-stream
          title: Media File
          type: string
        name:
          title: Name
          type: string
      required:
      - media_file
      - name
      title: Body_import_document_api_v1_documents_import__post
      type: object
...
    DocumentUpdateRequest:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      title: DocumentUpdateRequest
      type: object
...
    ExportFormat:
      enum:
      - VTT
      - SRT
      title: ExportFormat
      type: string
    ExportResult:
      properties:
        result:
          title: Result
          type: string
      required:
      - result
      title: ExportResult
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
  securitySchemes:
    HTTPBasic:
      scheme: basic
      type: http
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/documents/:
    get:
      operationId: list_documents_api_v1_documents__get
      parameters:
      - in: header
        name: authorization
        required: true
        schema:
          title: Authorization
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ApiDocumentWithTasks'
                title: Response List Documents Api V1 Documents  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Documents
    post:
      operationId: create_document_api_v1_documents__post
      parameters:
      - in: header
        name: authorization
        required: true
        schema:
          title: Authorization
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_create_document_api_v1_documents__post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocumentWithTasks'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Document
  /api/v1/documents/import/:
    post:
      operationId: import_document_api_v1_documents_import__post
      parameters:
      - in: header
        name: authorization
        required: true
        schema:
          title: Authorization
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_import_document_api_v1_documents_import__post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocumentWithTasks'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Import Document
  /api/v1/documents/{document_id}/:
    delete:
      operationId: delete_document_api_v1_documents__document_id___delete
      parameters:
      - in: path
        name: document_id
        required: true
        schema:
          format: uuid
          title: Document Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: Share-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Share-Token
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Document
    get:
      operationId: get_document_api_v1_documents__document_id___get
      parameters:
      - in: path
        name: document_id
        required: true
        schema:
          format: uuid
          title: Document Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: Share-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Share-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentWithAccessInfo'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Document
    patch:
      operationId: update_document_api_v1_documents__document_id___patch
      parameters:
      - in: path
        name: document_id
        required: true
        schema:
          format: uuid
          title: Document Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: Share-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Share-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentUpdateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocumentWithTasks'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Document
  /api/v1/documents/{document_id}/add_export_result/:
    post:
      operationId: add_export_result_api_v1_documents__document_id__add_export_result__post
      parameters:
      - in: path
        name: document_id
        required: true
        schema:
          format: uuid
          title: Document Id
          type: string
      - in: query
        name: task_id
        required: true
        schema:
          title: Task Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: Share-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Share-Token
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/ExportResult'
              - $ref: '#/components/schemas/ExportError'
              title: Result
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Add Export Result
  /api/v1/documents/{document_id}/add_media_file/:
    post:
      operationId: add_media_file_api_v1_documents__document_id__add_media_file__post
      parameters:
      - in: path
        name: document_id
        required: true
        schema:
          format: uuid
          title: Document Id
          type: string
      - in: header
        name: authorization
        required: true
        schema:
          title: Authorization
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_add_media_file_api_v1_documents__document_id__add_media_file__post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocumentWithTasks'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Add Media File
  /api/v1/documents/{document_id}/export/:
    get:
      operationId: export_api_v1_documents__document_id__export__get
      parameters:
      - in: path
        name: document_id
        required: true
        schema:
          format: uuid
          title: Document Id
          type: string
      - in: query
        name: format
        required: true
        schema:
          $ref: '#/components/schemas/ExportFormat'
      - in: query
        name: include_speaker_names
        required: true
        schema:
          title: Include Speaker Names
          type: boolean
      - in: query
        name: include_word_timing
        required: true
        schema:
          title: Include Word Timing
          type: boolean
      - in: query
        name: max_line_length
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Line Length
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: Share-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Share-Token
      responses:
        '200':
          content:
            text/plain:
              schema:
                type: string
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Export
  /api/v1/documents/{document_id}/media_files/:
    get:
      operationId: get_document_media_api_v1_documents__document_id__media_files__get
      parameters:
      - in: path
        name: document_id
        required: true
        schema:
          format: uuid
          title: Document Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: Share-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Share-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RemoteDocumentMedia'
                title: Response Get Document Media Api V1 Documents  Document Id  Media
                  Files  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Document Media
  /api/v1/documents/{document_id}/set_duration/:
    post:
      operationId: set_duration_api_v1_documents__document_id__set_duration__post
      parameters:
      - in: path
        name: document_id
        required: true
        schema:
          format: uuid
          title: Document Id
          type: string
      - in: header
        name: authorization
        required: true
        schema:
          title: Authorization
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetDurationRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocumentWithTasks'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Set Duration
...
  /api/v1/documents/{document_id}/share_tokens/{token_id}/:
    delete:
      operationId: delete_share_tokens_api_v1_documents__document_id__share_tokens__token_id___delete
      parameters:
      - in: path
        name: token_id
        required: true
        schema:
          format: uuid
          title: Token Id
          type: string
      - in: path
        name: document_id
        required: true
        schema:
          format: uuid
          title: Document Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: Share-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Share-Token
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Share Tokens
...